+++ CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file. Dates are d
 
 #### [Unreleased](https://github.com/henriquehbr/eswatch/compare/v0.11.0...HEAD)
 
-- chore: bump dependencies [`69a1168`](https://github.com/henriquehbr/eswatch/commit/69a1168432c3fb8811515488b6f3177499d8b2d1)
+- refactor(parseOptions): remove --splitting validation [`ea35cd6`](https://github.com/henriquehbr/eswatch/commit/ea35cd65ebdde0ce09e41363971861f12f8e9041)
+- chore: bump dependencies [`1304a76`](https://github.com/henriquehbr/eswatch/commit/1304a76bf9b6c1ff8a71e37561b8d0a83a940574)
 
 #### [v0.11.0](https://github.com/henriquehbr/eswatch/compare/v0.10.4...v0.11.0)